# docs/api-desktop-openapi.yaml
openapi: 3.1.0
...
info:
  title: heron desktop API
  version: 2026-04-25
  summary: Localhost HTTP control surface for the heron desktop daemon.
  description: |
    `herond` exposes meeting capture, transcripts, summaries, calendar
    reads, and a realtime event stream over HTTP on localhost. It is
    one of several transport projections of the same canonical trait
    surface; in-process Rust subscribers, Tauri IPC, MCP, and outbound
    webhooks are the others (`api-design-spec.md` §10, Invariant 13).

    External consumers (Raycast extensions, scripts, MCP wrappers, the
    Tauri shell when it later moves out-of-process) talk to this API.
    The Tauri shell currently uses Tauri IPC directly; HTTP is for
    *external* consumers.

    **Network posture: localhost-only.** v1 binds `127.0.0.1:7384`
    exclusively. A networked mode (with explicit pairing, short-lived
    tokens or mTLS) is deferred to v1.1; a single long-lived bearer
    in `~/.heron/cli-token` is not a defensible auth model over a
    network and v1 declines to ship one.

    **CORS: denied by default.** Browser-origin `fetch()` is rejected
    at the daemon. Browser-style consumers must use a non-`fetch`
    transport (Tauri's `invoke`, native HTTP from a desktop app, or a
    same-origin proxy). This avoids accidentally exposing the daemon
    to a malicious page on the same machine.

    **Versioning:** every request may include
    `Heron-API-Version: 2026-04-25`. The header is honored on every
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /meetings:
    get:
      tags: [meetings]
      summary: List meetings, newest first.
      parameters:
        - in: query
          name: since
          schema: { type: string, format: date-time }
          description: Filter to meetings created at or after this timestamp.
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/MeetingStatus" }
        - in: query
          name: platform
          schema: { $ref: "#/components/schemas/Platform" }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
          description: |
            Opaque cursor returned by a previous response in
            `next_cursor`. Cursor-based, not offset-based, so deletes
            don't shift the page.
      responses:
        "200":
          description: A page of meetings.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Meeting" }
                  next_cursor:
                    type: [string, "null"]
        "401": { $ref: "#/components/responses/Unauthorized" }

    post:
      tags: [meetings]
      summary: Manually start a meeting capture (escape hatch).
      description: |
        The happy path is ambient detection — `heron-session` (planned
        hub, today split between `heron-cli::session` and `heron-zoom`)
        emits `meeting.detected` and the daemon arms capture without a
        client ever calling this. This endpoint exists so an agent can
        force a capture when detection fails. If a client calls this in
        the steady state, detection is broken; treat it as a bug
        report, not a feature.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform: { $ref: "#/components/schemas/Platform" }
                hint:
                  type: string
                  description: |
                    Free-form hint forwarded to the orchestrator
                    (e.g. window title, meeting URL). Not a primary
                    identifier.
      responses:
        "202":
          description: Capture starting; poll the returned meeting.
          headers:
            Location:
              schema: { type: string }
              description: Path to the created meeting resource.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meeting" }
        "409":
          description: A capture is already in progress for this platform.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: |
            A required capture permission (Core Audio, mic,
            accessibility) is missing. The error body's `details`
            names which permission. User-actionable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /meetings/{meeting_id}:
    parameters:
      - $ref: "#/components/parameters/MeetingIdPath"
    get:
      tags: [meetings]
      summary: Fetch a single meeting.
      responses:
        "200":
          description: Meeting found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meeting" }
        "404": { $ref: "#/components/responses/NotFound" }
  /meetings/{meeting_id}/end:
    parameters:
      - $ref: "#/components/parameters/MeetingIdPath"
    post:
      tags: [meetings]
      summary: Manually end an in-progress meeting (escape hatch).
      description: |
        As with `POST /meetings`, this is an escape hatch. Detection
        normally drives the `recording → ended → done` transition; if
        you find yourself reaching for this, detection failed.
      responses:
        "204": { description: Meeting transitioned to `ended`. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            Meeting is not in a state that accepts `end` (e.g. already
            terminal). Body's `details.current_state` reports where it
            actually is.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /meetings/{meeting_id}/pause:
    parameters:
      - $ref: "#/components/parameters/MeetingIdPath"
    post:
      tags: [meetings]
      summary: Pause an in-progress capture mid-session.
      description: |
        Tier 3 #16. Drives `recording → paused`. The daemon-side
        capture pipeline drops audio frames on the floor while paused,
        so the saved WAV (and downstream transcript / m4a) reflect
        only the un-paused portion of the call. WAV writers stay open
        — a subsequent `resume` picks back up without sealing the file.
      responses:
        "204": { description: Capture transitioned to `paused`. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            Meeting is not in a state that accepts `pause` (e.g. it is
            already paused, or never reached `recording`). Body's
            `details.current_state` reports where it actually is.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /meetings/{meeting_id}/resume:
    parameters:
      - $ref: "#/components/parameters/MeetingIdPath"
    post:
      tags: [meetings]
      summary: Resume a paused capture.
      description: |
        Tier 3 #16. Drives `paused → recording`. New audio frames flow
        back into the WAV writers and the AX speaker collector;
        envelopes resume publishing on `/events`.
      responses:
        "204": { description: Capture transitioned to `recording`. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            Meeting is not in a state that accepts `resume` (e.g. it
            was never paused). Body's `details.current_state` reports
            where it actually is.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /meetings/{meeting_id}/transcript:
    parameters:
      - $ref: "#/components/parameters/MeetingIdPath"
      - in: query
        name: format
        schema: { type: string, enum: [json, md, vtt], default: json }
        description: |
          Browser fallback when the client cannot set an `Accept`
          header. If both are present, `Accept` wins.
    get:
      tags: [transcripts]
      summary: Read the meeting transcript.
      description: |
        Returns the orchestrator's finalized segments. While the
        meeting is `recording`, the response carries finalized
        segments only and `Transcript.status` is `partial`; live
        partials remain SSE-only on `/events`. Once the meeting
        reaches `ended | done`, status flips to `complete`.

        Content negotiation: `Accept: application/json | text/markdown
        | text/vtt`. The `?format=` query is the browser fallback;
        `Accept` wins on conflict.
      responses:
        "200":
          description: |
            Transcript readable. The body's `status` reports whether
            segments are still being added (`partial`) or the
            transcript is sealed (`complete`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transcript" }
            text/markdown:
              schema: { type: string }
            text/vtt:
              schema: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
        "423":
          description: |
            Vault temporarily locked (e.g. iCloud Drive evicted the
            file or another writer holds the path). User-actionable;
            retry once the vault settles.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /meetings/{meeting_id}/summary:
    parameters:
      - $ref: "#/components/parameters/MeetingIdPath"
    get:
      tags: [summaries]
      summary: Read the meeting summary if ready.
      description: |
        Summaries are generated asynchronously after the meeting
        reaches a terminal state. Subscribe to `summary.ready` on
        `/events` rather than polling.
      responses:
        "200":
          description: Summary ready.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Summary" }
        "202":
          description: |
            Summary not yet generated. Prefer subscribing to
            `summary.ready` on `/events`; `Retry-After` is a soft
            hint only.
          headers:
            Retry-After:
              schema: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
        "423":
          description: |
            Vault temporarily locked (e.g. iCloud eviction). Retry
            once the vault path is readable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "424":
          description: |
            Upstream LLM provider failed. The error body names the
            provider in `details.provider` and uses
            `code: HERON_E_LLM_PROVIDER_FAILED`. A `daemon.error`
            event with the same code fires on `/events`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /meetings/{meeting_id}/audio:
    parameters:
      - $ref: "#/components/parameters/MeetingIdPath"
      - in: header
        name: Range
        schema: { type: string, example: "bytes=0-65535" }
        description: Standard HTTP byte-range request.
      - in: header
        name: If-None-Match
        schema: { type: string }
        description: Conditional GET against `ETag`.
    get:
      tags: [meetings]
      summary: Download the m4a audio sidecar.
      description: |
        Streams the m4a file inline. Supports byte-range requests for
        seek. The sidecar lives in the user's vault directory; the
        daemon proxies bytes rather than redirecting to a `file://`
        URL because vault paths may be inside cloud-sync directories
        the consumer cannot read directly.

        **Availability:** audio is only served when the meeting status
        is `ended | done`. A still-recording meeting returns
        `425 Too Early` because the m4a is being written and a
        partial download is indistinguishable from a complete file.
        Wait for `meeting.completed` on `/events` before fetching.
      responses:
        "200":
          description: Full audio stream.
          headers:
            Accept-Ranges:
              schema: { type: string, enum: [bytes] }
            ETag:
              schema: { type: string }
            Content-Length:
              schema: { type: integer }
          content:
            audio/mp4:
              schema: { type: string, format: binary }
        "206":
          description: Partial content (range request).
          headers:
            Content-Range:
              schema: { type: string, example: "bytes 0-65535/1048576" }
            ETag:
              schema: { type: string }
          content:
            audio/mp4:
              schema: { type: string, format: binary }
        "304":
          description: Not modified (matched `If-None-Match`).
        "404": { $ref: "#/components/responses/NotFound" }
        "416":
          description: Requested range not satisfiable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "423":
          description: |
            Vault temporarily locked (e.g. iCloud eviction). Retry
            once the vault settles.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "425":
          description: |
            Meeting is still recording. The m4a is in flight; a
            partial download would be indistinguishable from a
            complete file. Wait for `meeting.completed` on `/events`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /calendar/upcoming:
    get:
      tags: [calendar]
      summary: List upcoming calendar events.
      description: |
        Read-only slice over the EventKit bridge already wired into
        `heron-vault`. Returned events are correlated with prior
        meetings where attendees overlap, so a consumer can fetch the
        full pre-meeting briefing in one call.
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: A list of calendar events.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CalendarEvent" }
        "503":
          description: Calendar permission missing. User-actionable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /context:
    put:
      tags: [context]
      summary: Attach pre-meeting context to a calendar event.
      description: |
        Stores user-supplied context keyed to a calendar event. When
        capture starts and the calendar event matches, the context is
        consumed by the orchestrator and (in v2) injected into the
        agent's system prompt during `loading_persona → tts_warming`
        per spec §8. Idempotent: the latest PUT for a given
        `calendar_event_id` wins.

        `calendar_event_id` is an EventKit identifier — a resolver
        input (Invariant 4), so it lives in the request body, never
        in a URL path.

        Context size is capped at 16K tokens before downstream
        consumption (Invariant 10). Larger payloads are accepted here
        but summarized at consumption time.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreMeetingContextRequest" }
      responses:
        "204": { description: Context stored. }
        "404":
          description: Calendar event not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Validation error (e.g. attendee email malformed).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token written by `herond` to `~/.heron/cli-token` (mode
        0600) on first run. Single-token model; rotate by deleting the
        file and restarting the daemon.
  parameters:
    MeetingIdPath:
      in: path
      name: meeting_id
      required: true
      schema: { $ref: "#/components/schemas/MeetingId" }
  responses:
    Unauthorized:
      description: Bearer token missing or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    MeetingId:
      type: string
      pattern: '^mtg_[0-9a-f-]+$'
      description: |
        Stripe-style prefixed UUIDv7. Wire form
        `mtg_<lowercase-hyphenated-uuid>` (e.g.
        `mtg_550e8400-e29b-71d4-a716-446655440000`); time-ordered
        because the underlying UUID is v7. Spec §2 originally
        named "Crockford base32" as the encoding, but the
        `heron_types::prefixed_id!` macro emits hyphenated UUIDs
        and that is what the daemon serves. A future migration to
        base32 (for compactness) is a separate concerted change
        with backward-compat parsing — out of scope for v1.
...
    Platform:
      type: string
      enum: [zoom, google_meet, microsoft_teams, webex]
      description: |
        Native meeting client. **v1 actually serves only `zoom`**;
        the other values are reserved for v1.1+ (Google Meet / Teams
        once the WebRTC track-ID interception path lands; Webex
        pending an accessibility-tree survey). A v1 daemon emitting
        anything other than `zoom` is a bug — clients should still
        accept the wider enum for forward compatibility.
    MeetingStatus:
      type: string
      enum: [detected, armed, recording, paused, ended, done, failed]
      description: |
        v1 desktop meeting lifecycle from `heron-cli::session` and
        `architecture.md` — distinct from spec §3, which is the v2
        bot FSM. `done` is terminal-success (transcript + summary
        written + audio sidecar verified). `failed` is
        terminal-failure (capture aborted, transcript orphaned).
        Both terminal states emit a single `meeting.completed` event
        on `/events` with `data.outcome` carrying success vs failure
        per Invariant 9. `paused` is the Tier 3 #16 mid-session pause
        state — daemon drops capture frames on the floor; `resume`
        flips back to `recording`.
    Confidence:
      type: string
      enum: [high, low]
      description: |
        AXObserver-attributed turns are `high`; turns marked `them`
        without a real display name are `low` (per the README's
        ~70/30 quality promise).
    Participant:
      type: object
      required: [display_name, identifier_kind, is_user]
      properties:
        display_name:
          type: string
          description: |
            Real display name when known (Zoom AX tree, WebRTC track
            label). For low-confidence turns this is "them" or the
            literal string "me".
        identifier_kind:
          type: string
          enum: [ax_tree, webrtc_track, mic, fallback]
        is_user:
          type: boolean
          description: True when this participant is the local user.
    Meeting:
      type: object
      required: [id, status, platform, started_at]
      properties:
        id: { $ref: "#/components/schemas/MeetingId" }
        status: { $ref: "#/components/schemas/MeetingStatus" }
        platform: { $ref: "#/components/schemas/Platform" }
        title: { type: [string, "null"] }
        calendar_event_id:
          type: [string, "null"]
          description: |
            EventKit identifier when correlation succeeded; `null`
            when the meeting was captured ad-hoc. **Correlation
            only** — never a heron operational input. Use `MeetingId`
            for anything other than cross-referencing with the user's
            calendar (Invariant 4).
        started_at: { type: string, format: date-time }
        ended_at: { type: [string, "null"], format: date-time }
        duration_secs: { type: [integer, "null"] }
        participants:
          type: array
          items: { $ref: "#/components/schemas/Participant" }
        transcript_status:
          type: string
          enum: [pending, partial, complete, failed]
        summary_status:
          type: string
          enum: [pending, ready, failed]
        tags:
          type: array
          items: { type: string }
          default: []
          description: |
            LLM-inferred topic tags lifted from the note's
            `Frontmatter.tags`. Empty for active captures (no summary
            yet) and for any meeting whose summarizer omitted them.
            Order is the writer's order so a frontend can render a
            stable chip list. Older daemon builds may omit the field
            entirely; clients SHOULD treat a missing `tags` as `[]`.
    TranscriptSegment:
      type: object
      required: [speaker, text, start_secs, end_secs, is_final]
      properties:
        speaker: { $ref: "#/components/schemas/Participant" }
        text: { type: string }
        start_secs: { type: number }
        end_secs: { type: number }
        confidence: { $ref: "#/components/schemas/Confidence" }
        is_final:
          type: boolean
          description: |
            False segments are subject to revision in a later
            partial. Only final segments persist to the vault.
    Transcript:
      type: object
      required: [meeting_id, status, segments]
      properties:
        meeting_id: { $ref: "#/components/schemas/MeetingId" }
        status:
          type: string
          enum: [pending, partial, complete, failed]
          description: |
            Lifecycle of the transcript itself, decoupled from the
            parent meeting. `partial` means more segments are still
            being added — subscribe to `/events` for live deltas.
            `complete` means the transcript is sealed.
        language:
          type: string
          example: en
        segments:
          type: array
          items: { $ref: "#/components/schemas/TranscriptSegment" }
...
    Summary:
      type: object
      required: [meeting_id, generated_at, text]
      properties:
        meeting_id: { $ref: "#/components/schemas/MeetingId" }
        generated_at: { type: string, format: date-time }
        text:
          type: string
          description: Markdown body, the same text written into the vault note.
        action_items:
          type: array
          items: { $ref: "#/components/schemas/ActionItem" }
        llm_provider: { type: string, example: anthropic }
        llm_model: { type: string, example: claude-opus-4-7 }
    CalendarEvent:
      type: object
      required: [id, title, start, end]
      properties:
        id:
          type: string
          description: |
            EventKit identifier. Resolver-input shape; not a heron
            primary key.
        title: { type: string }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        attendees:
          type: array
          items: { $ref: "#/components/schemas/AttendeeContext" }
        meeting_url: { type: [string, "null"], format: uri }
        related_meetings:
          type: array
          items: { $ref: "#/components/schemas/MeetingId" }
          description: |
            Prior captured meetings whose attendees overlap. Useful
            for clients building a pre-meeting briefing on top of the
            calendar event.
...
    PreMeetingContextRequest:
      type: object
      description: |
        Body shape for `PUT /context`. `calendar_event_id` is a
        resolver input (Invariant 4), so it lives in the body — never
        in a URL path.
      required: [calendar_event_id, context]
      properties:
        calendar_event_id:
          type: string
          description: |
            EventKit identifier of the calendar event the context
            attaches to.
        context: { $ref: "#/components/schemas/PreMeetingContext" }
...
    Error:
      type: object
      description: |
        MeetingBaaS-shape error envelope, adopted in spec §11 as
        heron's internal taxonomy projected over HTTP.
      required: [success, error, message, code, statusCode]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          description: PascalCase machine-readable error name.
          example: VaultLocked
        message:
          type: string
          description: Human-readable explanation. May change between versions.
        code:
          type: string
          pattern: '^HERON_E_[A-Z0-9_]+$'
          example: HERON_E_VAULT_LOCKED
        statusCode:
          type: integer
          example: 423
        details:
          type: object
          additionalProperties: true